requirements.txt
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyYAML==6.0.1

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Storage
minio==7.2.0
boto3==1.34.0

# ML & Embeddings
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
faiss-cpu==1.7.4

# NLP
langchain==0.1.0
langchain-community==0.0.10
pypdf==3.17.0
python-docx==1.1.0

# Spark
pyspark==3.5.0
pyarrow==14.0.1

# API
fastapi==0.109.0
uvicorn[standard]==0.25.0
gradio==4.12.0
python-multipart==0.0.6

# Utilities
aiohttp==3.9.1
httpx==0.26.0
tenacity==8.2.3
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1